Hide a bunch of classes
Make a bunch of code internal so it doesn't pollute namespaces of consumers. We may bring them back if there's good reason.
