+++ src/PostHog.AspNetCore/Registration.cs
@@ -1,10 +1,10 @@
 ﻿using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using PostHog.Api;
 using PostHog.Cache;
 using PostHog.Config;
 using PostHog.Library;
+using static PostHog.Library.Ensure;
 
 namespace PostHog;
 
@@ -18,7 +18,7 @@ public static class Registration
     /// </summary>
     /// <param name="builder">The <see cref="IHostApplicationBuilder"/>.</param>
     /// <returns>The passed in <see cref="IHostApplicationBuilder"/>.</returns>
-    public static IHostApplicationBuilder AddPostHog(this IHostApplicationBuilder builder)
+    public static IHttpClientBuilder AddPostHog(this IHostApplicationBuilder builder)
         => builder.AddPostHog(DefaultConfigurationSectionName);
 
     /// <summary>
@@ -28,9 +28,10 @@ public static class Registration
     /// <param name="builder">The <see cref="IHostApplicationBuilder"/>.</param>
     /// <param name="configurationSectionName">The configuration section name to grab PostHog options from.</param>
     /// <returns>The passed in <see cref="IHostApplicationBuilder"/>.</returns>
-    public static IHostApplicationBuilder AddPostHog(this IHostApplicationBuilder builder, string configurationSectionName)
-        => builder.AddPostHog(
-            (builder ?? throw new ArgumentNullException(nameof(builder))).Configuration.GetSection(configurationSectionName));
+    public static IHttpClientBuilder AddPostHog(
+        this IHostApplicationBuilder builder,
+        string configurationSectionName)
+        => builder.AddPostHog(NotNull(builder).Configuration.GetSection(configurationSectionName));
 
     /// <summary>
     /// Registers <see cref="PostHogClient"/> as a singleton. Looks for client configuration in the supplied
@@ -40,7 +41,7 @@ public static class Registration
     /// <param name="configurationSection">The configuration section where to load config settings.</param>
     /// <returns>The passed in <see cref="IHostApplicationBuilder"/>.</returns>
     /// <exception cref="ArgumentNullException">If <see cref="builder"/> is null.</exception>
-    public static IHostApplicationBuilder AddPostHog(
+    public static IHttpClientBuilder AddPostHog(
         this IHostApplicationBuilder builder,
         IConfigurationSection configurationSection)
     {
@@ -49,9 +50,7 @@ public static class Registration
         builder.Services.Configure<PostHogOptions>(configurationSection);
         builder.Services.AddSingleton<IFeatureFlagCache, HttpContextFeatureFlagCache>();
         builder.Services.AddSingleton<ITaskScheduler, TaskRunTaskScheduler>();
-        builder.Services.AddSingleton<IPostHogApiClient, PostHogApiClient>();
         builder.Services.AddSingleton<IPostHogClient, PostHogClient>();
-
-        return builder;
+        return builder.Services.AddHttpClient(nameof(PostHogClient));
     }
 }
