+++ src/PostHog/Api/PostHogApiClientExtensions.cs
@@ -19,7 +19,7 @@ internal static class PostHogApiClientExtensions
     /// <param name="cancellationToken">The cancellation token that can be used to cancel the operation.</param>
     /// <returns>An <see cref="ApiResult"/> with the result of the operation.</returns>
     public static async Task<ApiResult> IdentifyPersonAsync(
-        this IPostHogApiClient client,
+        this PostHogApiClient client,
         string distinctId,
         Dictionary<string, object>? userPropertiesToSet,
         Dictionary<string, object>? userPropertiesToSetOnce,
@@ -48,7 +48,7 @@ internal static class PostHogApiClientExtensions
     /// Identify a group with additional properties
     /// </summary>
     public static async Task<ApiResult> IdentifyGroupAsync(
-        this IPostHogApiClient client,
+        this PostHogApiClient client,
         string type,
         StringOrValue<int> key,
         Dictionary<string, object>? properties,
@@ -81,7 +81,7 @@ internal static class PostHogApiClientExtensions
     /// <param name="cancellationToken">The cancellation token that can be used to cancel the operation.</param>
     /// <returns>An <see cref="ApiResult"/> with the result of the operation.</returns>
     public static async Task<ApiResult> AliasAsync(
-        this IPostHogApiClient client,
+        this PostHogApiClient client,
         string distinctId,
         string alias,
         CancellationToken cancellationToken)
@@ -101,7 +101,7 @@ internal static class PostHogApiClientExtensions
     /// Unlink future events with the current user. Call this when a user logs out.
     /// </summary>
     public static async Task ResetAsync(
-        this IPostHogApiClient client,
+        this PostHogApiClient client,
         string distinctId,
         CancellationToken cancellationToken)
     {
@@ -116,7 +116,7 @@ internal static class PostHogApiClientExtensions
     /// Capture an event with optional properties
     /// </summary>
     public static async Task<ApiResult> CaptureAsync(
-        this IPostHogApiClient client,
+        this PostHogApiClient client,
         string distinctId,
         string eventName,
         Dictionary<string, object>? properties,
@@ -135,12 +135,12 @@ internal static class PostHogApiClientExtensions
     /// Capture an event with optional properties
     /// </summary>
     public static async Task<ApiResult> CaptureAsync(
-        this IPostHogApiClient client,
+        this PostHogApiClient client,
         string distinctId,
         string eventName) => await client.CaptureAsync(distinctId, eventName, null, CancellationToken.None);
 
     static async Task<ApiResult> SendEventAsync(
-        this IPostHogApiClient client,
+        this PostHogApiClient client,
         string distinctId,
         string eventName,
         Dictionary<string, object>? properties,
