+++ src/PostHog/Library/TimeProviderSystemClock.cs
@@ -6,7 +6,7 @@ namespace PostHog.Library;
 /// An implementation of <see cref="ISystemClock"/> that uses a
 /// <see cref="TimeProvider"/> under the hood.
 /// </summary>
-public class TimeProviderSystemClock(TimeProvider timeProvider) : ISystemClock
+internal class TimeProviderSystemClock(TimeProvider timeProvider) : ISystemClock
 {
     /// <summary>
     /// Retrieve's the current system's time in UTC.
