+++ src/PostHog/PostHog.csproj
@@ -8,6 +8,7 @@
 
     <ItemGroup>
         <PackageReference Include="System.Text.Json" Version="9.0.0" />
+        <PackageReference Include="Microsoft.Extensions.Http" Version="9.0.0" />
         <PackageReference Include="Microsoft.Extensions.Logging" Version="9.0.0" />
         <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="9.0.1" />
     </ItemGroup>
